# Remove MapSector, take 2