Retry of #10656, should make more sense after #12321

This completely removes MapSector and stores all blocks directly on the Map (as unordered_map<v3s16, unique_ptr<MapBlock>>, similar to MapSector + Map before). It also gets rid of the last seen block caches we had in Map and MapSector. They did not seem to do much, but I am happy to bring a cache back.

This makes the code cleaner. It's somewhat hard to say whether the one call to to the new `Map::getBlockNoCreateNoEx` takes more time than `MapSector::getBlockNoCreateNoEx` + `Map::getSectorNoGenerateNoLock`. Maybe a little bit, but it seems within the noise. I could use some help testing.

- Goal of the PR
See above, code cleanliness

- How does the PR work?
See above.

- Does it resolve any reported issue?
Nope.

- Does this relate to a goal in [the roadmap](https://github.com/luanti-org/luanti/blob/master/doc/direction.md)?
- If not a bug fix, why is this PR needed? What usecases does it solve?

## To do

This PR is Ready for Review.

It's best to review this without white space, as a lot of indentation has changed.

## How to test

Try any world. Large viewing-ranges, many objects, ABMs/LBMs, etc. Everything should still work without measurable performance degradation.
Help with perf-testing...