@sfan5 2025-06-10
Writing a new benchmark that simulates typical "block lookup" workloads is probably needed. I'll try to take a look in a few days.