@lhofhansl 2025-06-11
The benchmark
1. creates n^3 blocks.
2. Reads n^3 blocks from empty map.
3. Reads n^3 blocks from a filled map.

Without PR:

```
/home/lars/dev/luanti/luanti/src/benchmark/benchmark_map.cpp:60
...............................................................................

benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
create_10                                      100             2 2297.200000 us 
                                     1049.975255 us 1047.560140 us 1056.065045 us
                                      17.940952 us   5.053570 us  32.463753 us
                                                                               
readEmpty_10                                   100             3 2114.100000 us
                                       7.069947 us   7.045453 us   7.135537 us
                                       0.198113 us   0.097999 us   0.381687 us
                                                                               
readFilled_10                                  100             1 4042.600000 us
                                      42.957440 us  42.786930 us  43.236350 us
                                       1.088977 us   0.752809 us   1.595234 us
                                                                               
create_40                                      100             1 39194051.600000 us  
                                     130371.058440 us 129560.436490 us 131607.996860 us
                                     5031.845576 us 3652.162049 us 6961.144284 us
                                                                               
readEmpty_40                                   100             1 35050.700000 us
                                     350.932650 us 350.476130 us 352.642040 us
                                       4.034970 us   0.973571 us   9.391777 us
                                                                               
readFilled_40                                  100             1 335724.300000 us
                                     5967.241480 us 5947.937560 us 5994.263980 us
                                     115.687428 us  88.594537 us 178.529466 us


```

With PR:
```
/home/lars/dev/luanti/luanti/src/benchmark/benchmark_map.cpp:60
...............................................................................

benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
create_10                                      100             1 198922.400000 us 
                                     2097.784720 us 2084.694210 us 2119.881760 us 
                                      84.917404 us  58.873491 us 154.438321 us 
                                                                               
readEmpty_10                                   100             3 2225.100000 us 
                                       7.412543 us   7.388853 us   7.486963 us 
                                       0.180890 us   0.005467 us   0.401566 us 
                                                                               
readFilled_10                                  100             1 3228.900000 us 
                                      41.242620 us  41.127520 us  41.442860 us 
                                       0.756750 us   0.458286 us   1.223904 us 
                                                                               
create_40                                      100             1 40961703.900000 us 
                                     134288.132560 us 133316.909190 us 135524.409150 us 
                                     5565.385515 us 4446.704055 us 7393.146082 us 
                                                                               
readEmpty_40                                   100             1 36905.000000 us 
                                     367.876550 us 367.406200 us 369.180160 us 
                                       3.685272 us   1.698806 us   7.882470 us 
                                                                               
readFilled_40                                  100             1 328362.000000 us 
                                     4940.859590 us 4870.669540 us 5055.478410 us 
                                     448.001961 us 314.352094 us 797.052792 us 
                                                                               
                                  1131.086165 us 899.606048 us 1401.928772 us
```