@lhofhansl 2025-06-11
So it's initial insert that slows down (inserting the first 1000 blocks is almost 50% slower, but then it amortizes and reading is faster!