@lhofhansl 2025-06-11
Without PR:
```
/home/lars/dev/luanti/luanti/src/benchmark/benchmark_map.cpp:60
...............................................................................

benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
create_10                                      100             2 2584.200000 us 
                                     1054.627360 us 1050.298225 us 1070.585490 us 
                                      37.199414 us   6.667865 us  85.338844 us 
                                                                               
readEmpty_10                                   100             3 2113.200000 us 
                                       7.058887 us   7.036193 us   7.125680 us 
                                       0.182100 us   0.067950 us   0.381318 us 
                                                                               
readFilled_10                                  100             1 4016.400000 us 
                                      45.779640 us  42.589340 us  61.198310 us 
                                      30.779771 us   1.195012 us  73.374561 us 
                                                                               
create_40                                      100             1 40445260.000000 us 
                                     133039.292570 us 131814.715650 us 134739.393990 us 
                                     7289.535216 us 5682.957367 us 9489.607662 us 
                                                                               
readEmpty_40                                   100             1 36217.100000 us 
                                     362.897190 us 362.456770 us 363.680360 us 
                                       2.908522 us   1.902607 us   5.048469 us 
                                                                               
readFilled_40                                  100             1 334913.500000 us 
                                     6209.366180 us 6134.386520 us 6327.388770 us 
                                     473.601472 us 346.974037 us 714.828614 us 
                                                                               
```

With PR:
```
/home/lars/dev/luanti/luanti/src/benchmark/benchmark_map.cpp:60
...............................................................................

benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
create_10                                      100             1 199117.100000 us 
                                     2095.479630 us 2083.275560 us 2109.655180 us 
                                      66.830982 us  57.799232 us  88.855737 us 
                                                                               
readEmpty_10                                   100             3 2220.900000 us 
                                       7.428383 us   7.395187 us   7.511433 us 
                                       0.241655 us   0.066126 us   0.428742 us 
                                                                               
readFilled_10                                  100             1 3173.400000 us 
                                      41.430870 us  41.261390 us  41.671040 us 
                                       1.021421 us   0.756137 us   1.361186 us 
                                                                               
create_40                                      100             1 39802090.700000 us 
                                     132191.186410 us 131461.280770 us 132909.891380 us 
                                     3715.792304 us 3455.360010 us 4027.549907 us 
                                                                               
readEmpty_40                                   100             1 36717.000000 us 
                                     367.803630 us 367.318940 us 369.335920 us 
                                       3.980749 us   1.518534 us   8.813921 us 
                                                                               
readFilled_40                                  100             1 326208.100000 us 
                                     5028.328630 us 4952.833950 us 5137.409770 us 
                                     459.210272 us 341.455438 us 667.791822 us 
```