@lhofhansl 2025-06-11
Updated the benchmark... Numbers:

Without PR:
```
/home/lars/dev/luanti/luanti/src/benchmark/benchmark_map.cpp:60
...............................................................................

benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev

create_10                                      100             2 1805.200000 us
                                     792.141665 us 790.640850 us 797.487635 us
                                      12.540803 us   2.451009 us  28.217173 us

readEmpty_10                                   100             4 2118.400000 us
                                       5.310515 us   5.286767 us   5.369988 us
                                       0.173357 us   0.049283 us   0.306930 us

readFilled_10                                  100             1 2964.600000 us
                                      31.845490 us  31.703500 us  32.085470 us
                                       0.914539 us   0.594436 us   1.262416 us

readEmptyYCol_10                               100             4 1819.600000 us
                                       4.562655 us   4.547658 us   4.591605 us
                                       0.104242 us   0.061549 us   0.155485 us

readFilledYCol_10                              100             1 1976.300000 us
                                      25.693110 us  25.510060 us  25.971010 us
                                       1.136008 us   0.812410 us   1.537206 us 
                                                                              
readEmptyNodes_10                              100             1 130412.400000 us
                                     1304.707710 us 1304.069220 us 1305.376270 us
                                       3.340101 us   2.932630 us   4.063785 us 
                                                                              
readFilledNodes_10                             100             1 177107.900000 us
                                     1785.441420 us 1782.371760 us 1790.354860 us
                                      19.445071 us  13.243015 us  30.500917 us 
                                                                               
create_40                                      100             1 37368745.900000 us
                                     122949.868560 us 121404.066420 us 124838.473260 us
                                     8725.638573 us 7372.950501 us 9861.471032 us
                                                                               
readEmpty_40                                   100             1 32555.900000 us
                                     326.103110 us 325.631900 us 327.823600 us
                                       4.083394 us   1.124926 us   9.385779 us 
                                                                               
readFilled_40                                  100             1 295701.400000 us
                                     5457.360160 us 5430.097550 us 5492.168140 us
                                     156.233191 us 125.200763 us 192.947423 us 
                                                                              
readEmptyYCol_40                               100             1 31622.700000 us
                                     316.843670 us 316.356240 us 318.451750 us
                                       4.029286 us   1.155122 us   8.956497 us 
                                                                               
readFilledYCol_40                              100             1 199250.700000 us
                                     3968.897300 us 3918.391060 us 4025.714650 us
                                     271.072781 us 223.823701 us 342.472614 us 
                                                                              
readEmptyNodes_40                              100             1 8341849.100000 us
                                     83456.892110 us 83448.204440 us 83467.134600 us
                                      47.942549 us  40.441283 us  59.254732 us 
                                                                               
readFilledNodes_40                             100             1 12782143.100000 us
                                     127507.644080 us 127348.643470 us 127665.387350 us
                                     804.850989 us 743.446247 us 885.410283 us  

```

With PR:
```
/home/lars/dev/luanti/luanti/src/benchmark/benchmark_map.cpp:60
...............................................................................

benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev

create_10                                      100             1 154705.300000 us
                                     1568.149410 us 1563.320720 us 1573.425570 us
                                      25.796026 us  23.837122 us  31.446003 us

readEmpty_10                                   100             3 1667.100000 us
                                       5.598507 us   5.566753 us   5.660443 us
                                       0.218593 us   0.127652 us   0.335988 us

readFilled_10                                  100             1 2295.200000 us
                                      27.538160 us  27.364160 us  27.851080 us
                                       1.156808 us   0.753453 us   1.847042 us

readEmptyYCol_10                               100             4 1914.800000 us
                                       4.826585 us   4.787177 us   4.925163 us
                                       0.286256 us   0.073285 us   0.517193 us

readFilledYCol_10                              100             1 2413.100000 us
                                      30.101520 us  29.975600 us  30.304790 us
                                       0.798881 us   0.549386 us   1.183209 us

readEmptyNodes_10                              100             1 128242.600000 us
                                     1282.253720 us 1281.798410 us 1282.804880 us
                                       2.538217 us   2.146162 us   3.580772 us

readFilledNodes_10                             100             1 314586.500000 us
                                     3151.777090 us 3150.919170 us 3153.086680 us
                                       5.332024 us   3.850324 us   8.132179 us

create_40                                      100             1 36891805.200000 us
                                     123629.766690 us 122767.145140 us 124809.957790 us
                                     5129.900009 us 3886.354034 us 7316.349703 us

readEmpty_40                                   100             1 34096.300000 us
                                     341.609940 us 341.228940 us 342.619820 us
                                       2.943417 us   1.363743 us   6.212667 us

readFilled_40                                  100             1 290564.200000 us
                                     4525.552100 us 4486.235680 us 4570.621790 us
                                     215.119800 us 189.156973 us 253.551072 us

readEmptyYCol_40                               100             1 33275.700000 us
                                     332.994630 us 332.712900 us 333.867010 us
                                       2.291949 us   0.946071 us   5.055097 us

readFilledYCol_40                              100             1 463596.500000 us
                                     7202.239250 us 7118.657030 us 7315.638800 us
                                     493.650160 us 384.958537 us 653.962477 us

readEmptyNodes_40                              100             1 8203101.000000 us
                                     82034.667750 us 82021.791010 us 82053.102130 us
                                      77.693298 us  58.701021 us 124.361061 us

readFilledNodes_40                             100             1 21193258.300000 us
                                     213052.205990 us 212720.094350 us 213525.942270 us
                                     2009.248138 us 1522.096589 us 2615.970734 us


```